Причесывает код
