+++ src/components/favorites/favorites.test.js
@@ -1,15 +1,18 @@
 
+import configureStore from "redux-mock-store";
+import Favorites from "./favorites.jsx";
+import NameSpace from "../../reducer/name-space.js";
 import React from "react";
 import renderer from "react-test-renderer";
-import Favorites from "./favorites.jsx";
-import configureStore from "redux-mock-store";
-import {Provider} from "react-redux";
-import NameSpace from "../../reducer/name-space.js";
-import {Operation} from '../../reducer/data/data';
 import thunk from 'redux-thunk';
 import {BrowserRouter} from "react-router-dom";
+import {Operation} from '../../reducer/data/data';
+import {Provider} from "react-redux";
+
+
 const middlewares = [thunk];
 const mockStore = configureStore(middlewares);
+
 jest.mock(`../../reducer/data/data`);
 Operation.loadFavoriteOffers = () => (dispatch) => dispatch(jest.fn());
 Operation.changeFavoriteStatus = () => (dispatch) => dispatch(jest.fn());
@@ -49,25 +52,6 @@ const offers = [
       isSuper: true,
     },
     coords: [52.39874984984841, 4.82456445558843],
-    reviews: [
-      {
-        reviewId: 14,
-        reviewName: `Ozzy`,
-        reviewAvatar: `src`,
-        reviewGrade: 4,
-        reviewText: `Id minim labore ut velit sit velit.Magna deserunt reprehenderit consequat elit cupidatat proident nostrud amet minim nulla.`,
-        reviewTime: `May 8, 2016`,
-      },
-      {
-        reviewId: 22,
-        reviewName: `Mick`,
-        reviewAvatar: `src`,
-        reviewGrade: 2,
-        reviewText: `Incididunt fugiat non aliqua eu nisi.Id Lorem cillum non voluptate nulla id fugiat Lorem exercitation irure ullamco enim veniam ullamco.`,
-        reviewTime: `June 5, 2018`,
-      },
-    ],
-    nearOffers: [1]
   },
 ];
 
@@ -82,15 +66,20 @@ describe(`Favorites`, () => {
         userEmail: ``,
       }
     });
+
+    const onLoadFavoriteOffers = jest.fn();
+    const changeFavoriteStatus = jest.fn();
+    const onCardHeaderClick = jest.fn();
+
     const tree = renderer
         .create(
             <Provider store={store}>
               <BrowserRouter>
                 <Favorites
                   authorizationStatus={`AUTH`}
-                  onLoadFavoriteOffers={() => {}}
-                  changeFavoriteStatus={() => {}}
-                  onCardHeaderClick={() => {}}
+                  onLoadFavoriteOffers={onLoadFavoriteOffers}
+                  changeFavoriteStatus={changeFavoriteStatus}
+                  onCardHeaderClick={onCardHeaderClick}
                 />
               </BrowserRouter>
             </Provider>, {createNodeMock: () => {
