+++ src/components/main-screen/main-screen.test.js
@@ -1,11 +1,11 @@
 
+import configureStore from "redux-mock-store";
+import MainScreen from "./main-screen.jsx";
+import NameSpace from "../../reducer/name-space.js";
 import React from "react";
 import renderer from "react-test-renderer";
-import MainScreen from "./main-screen.jsx";
-import configureStore from "redux-mock-store";
-import {Provider} from "react-redux";
-import NameSpace from "../../reducer/name-space.js";
 import {BrowserRouter} from "react-router-dom";
+import {Provider} from "react-redux";
 
 const mockStore = configureStore([]);
 
@@ -44,25 +44,6 @@ const offers = [
       isSuper: true,
     },
     coords: [52.39874984984841, 4.82456445558843],
-    reviews: [
-      {
-        reviewId: 14,
-        reviewName: `Ozzy`,
-        reviewAvatar: `src`,
-        reviewGrade: 4,
-        reviewText: `Id minim labore ut velit sit velit.Magna deserunt reprehenderit consequat elit cupidatat proident nostrud amet minim nulla.`,
-        reviewTime: `May 8, 2016`,
-      },
-      {
-        reviewId: 22,
-        reviewName: `Mick`,
-        reviewAvatar: `src`,
-        reviewGrade: 2,
-        reviewText: `Incididunt fugiat non aliqua eu nisi.Id Lorem cillum non voluptate nulla id fugiat Lorem exercitation irure ullamco enim veniam ullamco.`,
-        reviewTime: `June 5, 2018`,
-      },
-    ],
-    nearOffers: [1]
   },
 ];
 const cities = [
