+++ src/components/offer-card/offer-card.e2e.test.js
@@ -39,25 +39,6 @@ const userChoise = {
     isSuper: true,
   },
   coords: [52.39874984984841, 4.82456445558843],
-  reviews: [
-    {
-      reviewId: 14,
-      reviewName: `Ozzy`,
-      reviewAvatar: `src`,
-      reviewGrade: 4,
-      reviewText: `Id minim labore ut velit sit velit.Magna deserunt reprehenderit consequat elit cupidatat proident nostrud amet minim nulla.`,
-      reviewTime: `May 8, 2016`,
-    },
-    {
-      reviewId: 22,
-      reviewName: `Mick`,
-      reviewAvatar: `src`,
-      reviewGrade: 2,
-      reviewText: `Incididunt fugiat non aliqua eu nisi.Id Lorem cillum non voluptate nulla id fugiat Lorem exercitation irure ullamco enim veniam ullamco.`,
-      reviewTime: `June 5, 2018`,
-    },
-  ],
-  nearOffers: [1]
 };
 
 Enzyme.configure({
