+++ src/components/private-route/private-route.jsx
@@ -7,6 +7,7 @@ import {AuthorizationStatus} from "../../reducer/user/user.js";
 
 const PrivateRoute = (props) => {
   const {render, path, exact, authorizationStatus} = props;
+
   return (
     <Route
       path={path}
