+++ src/components/sing-in/sing-in.jsx
@@ -12,10 +12,10 @@ class SingIn extends PureComponent {
     this.loginRef = createRef();
     this.passwordRef = createRef();
 
-    this.handleSubmit = this.handleSubmit.bind(this);
+    this._handleSubmit = this._handleSubmit.bind(this);
   }
 
-  handleSubmit(evt) {
+  _handleSubmit(evt) {
     const {onSubmit} = this.props;
 
     evt.preventDefault();
@@ -58,7 +58,7 @@ class SingIn extends PureComponent {
           <div className="page__login-container container">
             <section className="login">
               <h1 className="login__title">Sign in</h1>
-              <form className="login__form form" action="#" method="post" onSubmit={this.handleSubmit}>
+              <form className="login__form form" action="#" method="post" onSubmit={this._handleSubmit}>
                 <div className="login__input-wrapper form__input-wrapper">
                   <label className="visually-hidden">E-mail</label>
                   <input className="login__input form__input" type="email" name="email" placeholder="Email" required=""
@@ -82,7 +82,6 @@ class SingIn extends PureComponent {
           </div>
         </main>
       </div>
-
     );
   }
 }
