+++ src/components/sort/sort.e2e.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import Enzyme, {mount} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-
 import {SortType} from '../../const.js';
 import {Sort} from './sort.jsx';
+
 
 Enzyme.configure({
   adapter: new Adapter(),
