+++ src/components/sort/sort.test.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-
 import {Sort} from './sort.jsx';
 
 it(`Should Sort render correctly`, () => {
